add utils submodule, rename functions
